Noise - Disintegration
Basic uniform has disintegration feature, adds a random colour splatter across a mesh based on a noise texture generated.

Can be enabled/disabled with a uniform bool, with parameters to define the alpha cutoffs of the filter.
Colour can be adjusted by uniform.

Results and effect is questionable in terms of aesthetic.